+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/model/InvoiceEditModel.java
@@ -30,7 +30,7 @@ public final class InvoiceEditModel extends SwingEntityEditModel {
 	public InvoiceEditModel(EntityConnectionProvider connectionProvider) {
 		super(Invoice.TYPE, connectionProvider);
 		persist(Invoice.CUSTOMER_FK).set(false);
-		valueEdited(Invoice.CUSTOMER_FK).addConsumer(this::setAddress);
+		edited(Invoice.CUSTOMER_FK).addConsumer(this::setAddress);
 	}
 
 	private void setAddress(Entity customer) {